+++ examples/basics/workflows/workflow-with-sequential-steps/index.ts
@@ -11,7 +11,7 @@ async function main() {
       doubledValue: z.number(),
     }),
     execute: async ({ context }) => {
-      const inputValue = context?.getStepPayload<{ inputValue: number }>('trigger')?.inputValue;
+      const inputValue = context?.getStepResult<{ inputValue: number }>('trigger')?.inputValue;
       if (!inputValue) throw new Error('No input value provided');
       const doubledValue = inputValue * 2;
       return { doubledValue };
@@ -27,9 +27,7 @@ async function main() {
       incrementedValue: z.number(),
     }),
     execute: async ({ context }) => {
-      const valueToIncrement = context?.getStepPayload<{ doubledValue: number }>(
-        'stepOne',
-      )?.doubledValue;
+      const valueToIncrement = context?.getStepResult<{ doubledValue: number }>('stepOne')?.doubledValue;
       if (!valueToIncrement) throw new Error('No value to increment provided');
       const incrementedValue = valueToIncrement + 1;
       return { incrementedValue };