+++ examples/quick-start/src/mastra/agents/agent.ts
@@ -1,4 +1,3 @@
-import { groq } from '@ai-sdk/groq';
 import { openai } from '@ai-sdk/openai';
 import { Agent } from '@mastra/core';
 
@@ -7,10 +6,4 @@ export const catOne = new Agent({
   instructions:
     'You are a feline expert with comprehensive knowledge of all cat species, from domestic breeds to wild big cats. As a lifelong cat specialist, you understand their behavior, biology, social structures, and evolutionary history in great depth. If you are asked for a specie name, do not return a paragraph, a succint two or three letter name of the species will suffice.',
   model: openai('gpt-4o'),
-});
-
-export const agentTwo = new Agent({
-  name: 'Agent Two',
-  instructions: 'Do this',
-  model: groq('llama3-groq-70b-8192-tool-use-preview'),
 });