+++ examples/quick-start/src/mastra/index.ts
@@ -1,12 +1,13 @@
 import { createLogger, Mastra } from '@mastra/core';
 
-import { catOne, agentTwo } from './agents/agent';
+import { catOne } from './agents/agent';
 import { logCatWorkflow } from './workflow';
 
 export const mastra = new Mastra({
-  agents: { catOne, agentTwo },
+  agents: { catOne },
   workflows: { logCatWorkflow },
   logger: createLogger({
+    name: 'Mastra',
     level: 'debug',
   }),
 });