+++ packages/core/src/agent/types.ts
@@ -8,6 +8,7 @@ import type { CoreMessage, OutputType } from '../llm';
 import type { MastraMemory } from '../memory/memory';
 import type { MemoryConfig } from '../memory/types';
 import type { ToolAction } from '../tools';
+import type { CompositeVoice } from '../voice';
 
 export type { Message as AiMessageType } from 'ai';
 
@@ -28,6 +29,7 @@ export interface AgentConfig<
   metrics?: TMetrics;
   evals?: TMetrics;
   memory?: MastraMemory;
+  voice?: CompositeVoice;
 }
 
 export interface AgentGenerateOptions<Z extends ZodSchema | JSONSchema7 | undefined = undefined> {
@@ -42,6 +44,7 @@ export interface AgentGenerateOptions<Z extends ZodSchema | JSONSchema7 | undefi
   output?: OutputType | Z;
   temperature?: number;
   toolChoice?: 'auto' | 'required';
+  experimental_output?: Z;
 }
 
 export interface AgentStreamOptions<Z extends ZodSchema | JSONSchema7 | undefined = undefined> {
@@ -57,4 +60,5 @@ export interface AgentStreamOptions<Z extends ZodSchema | JSONSchema7 | undefine
   output?: OutputType | Z;
   temperature?: number;
   toolChoice?: 'auto' | 'required';
+  experimental_output?: Z;
 }