+++ packages/create-mastra/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-mastra
 
+## 0.1.6-alpha.0
+
+### Patch Changes
+
+- 2d68431: Fix mastra server error processing
+
+## 0.1.5
+
+### Patch Changes
+
+- 967da43: Logger, transport fixes
+
 ## 0.1.4
 
 ### Patch Changes