+++ speech/playai/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @mastra/speech-playai
 
+## 0.1.4-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [6cb63e0]
+  - @mastra/core@0.4.2-alpha.1
+
+## 0.1.4-alpha.2
+
+### Patch Changes
+
+- 41d0166: deprecate @mastra/speech-playai for @mastra/voice-playai
+- Updated dependencies [7fceae1]
+- Updated dependencies [f626fbb]
+  - @mastra/core@0.4.2-alpha.0
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [ce44b9b]
+- Updated dependencies [967da43]
+- Updated dependencies [b405f08]
+  - @mastra/core@0.4.1
+
 ## 0.1.3
 
 ### Patch Changes